refactor: split capital calculation module
